Add `make localserver` target to run the CM server locally.
Factors the CGI-wrapping HTTPServer out of test_server.py into
centurymetadata/server/devserver.py so both the tests and the new
tools/localserver.py share it. `make localserver` spins up the API on
localhost:8199 (PORT=... to override) against a fresh tmpdir, letting
EXAMPLES.md's curl/centurytool steps be run against a real local server
instead of the production testapi.centurymetadata.org.
