+++ python/tests/test_server.py
@@ -1,13 +1,11 @@
 #! /usr/bin/env python3
 """Integration tests for the centurymetadata server and centurytool."""
 import json
-import os
 import random
 import subprocess
 import sys
 import threading
 import urllib.request
-from http.server import BaseHTTPRequestHandler, HTTPServer
 from pathlib import Path
 from typing import Dict, Generator, Set, Tuple
 
@@ -15,54 +13,9 @@ import pytest
 from secp256k1 import PrivateKey
 
 import centurymetadata
+from centurymetadata.server.devserver import CenturyHTTPServer, call_server
 
-SERVER_PY = Path(__file__).parent.parent / 'centurymetadata' / 'server' / 'server.py'
 TOOL_PY = Path(__file__).parent.parent.parent / 'examples' / 'centurytool.py'
-
-
-# ── CGI subprocess helper ─────────────────────────────────────────────────────
-
-def call_server(basedir: Path, method: str, path: str,
-                body: bytes = b'', content_type: str = '',
-                extra_env: Dict[str, str] = {}) -> Tuple[int, Dict[str, str], bytes]:
-    """Invoke server.py as a CGI subprocess and return (status, headers, body)."""
-    env = os.environ.copy()
-    env['REQUEST_METHOD'] = method
-    env['PATH_INFO'] = path
-    env['CENTURYMETADATA_BASEDIR'] = str(basedir)
-    if body:
-        env['CONTENT_LENGTH'] = str(len(body))
-    if content_type:
-        env['CONTENT_TYPE'] = content_type
-    env.update(extra_env)
-
-    result = subprocess.run(
-        [sys.executable, str(SERVER_PY)],
-        input=body,
-        capture_output=True,
-        env=env
-    )
-
-    if result.returncode != 0:
-        raise RuntimeError(f"Server subprocess failed (rc={result.returncode}):\n"
-                           f"{result.stderr.decode(errors='replace')}")
-
-    stdout = result.stdout
-    header_part, resp_body = stdout.split(b'\n\n', 1) if b'\n\n' in stdout else (stdout, b'')
-
-    status = 200
-    headers: Dict[str, str] = {}
-    for line in header_part.decode('utf-8', errors='replace').split('\n'):
-        line = line.strip()
-        if not line:
-            continue
-        if line.startswith('Status:'):
-            status = int(line.split()[1])
-        elif ':' in line:
-            k, v = line.split(':', 1)
-            headers[k.strip()] = v.strip()
-
-    return status, headers, resp_body
 
 
 # ── Fixtures ──────────────────────────────────────────────────────────────────
@@ -89,43 +42,9 @@ def keys() -> Dict:
     }
 
 
-class _CenturyHandler(BaseHTTPRequestHandler):
-    """Thin HTTP wrapper that delegates each request to server.py as a subprocess."""
-
-    server: '_CenturyHTTPServer'
-
-    def do_GET(self) -> None:
-        self._dispatch(b'')
-
-    def do_POST(self) -> None:
-        length = int(self.headers.get('Content-Length', 0))
-        self._dispatch(self.rfile.read(length))
-
-    def _dispatch(self, body: bytes) -> None:
-        content_type = self.headers.get('Content-Type', '')
-        status, headers, resp_body = call_server(
-            self.server.basedir, self.command, self.path, body, content_type
-        )
-        self.send_response(status)
-        for k, v in headers.items():
-            self.send_header(k, v)
-        self.send_header('Content-Length', str(len(resp_body)))
-        self.end_headers()
-        self.wfile.write(resp_body)
-
-    def log_message(self, *args: object) -> None:
-        pass  # suppress output during tests
-
-
-class _CenturyHTTPServer(HTTPServer):
-    def __init__(self, basedir: Path) -> None:
-        self.basedir = basedir
-        super().__init__(('localhost', 0), _CenturyHandler)
-
-
 @pytest.fixture()
 def http_server(basedir: Path) -> Generator[Tuple[str, Path], None, None]:
-    httpd = _CenturyHTTPServer(basedir)
+    httpd = CenturyHTTPServer(basedir)
     port = httpd.server_address[1]
     t = threading.Thread(target=httpd.serve_forever)
     t.daemon = True
