+++ tools/localserver.py
@@ -1,0 +1,45 @@
+#!/usr/bin/env python3
+"""Run a local centurymetadata server for manual testing.
+
+Creates the standard directory skeleton in a fresh tmpdir (unless
+--basedir is given) and serves the CGI API over plain HTTP until
+interrupted.
+
+Run from the python/ subdirectory (which has the venv and dependencies):
+
+    cd python && uv run python ../tools/localserver.py
+"""
+import argparse
+import shutil
+import tempfile
+from pathlib import Path
+
+from centurymetadata.server.devserver import CenturyHTTPServer
+
+
+def main() -> None:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--host", default="localhost")
+    parser.add_argument("--port", type=int, default=8199)
+    parser.add_argument("--basedir", type=Path, default=None,
+                        help="Storage directory (default: a fresh tmpdir, removed on exit)")
+    args = parser.parse_args()
+
+    basedir = args.basedir or Path(tempfile.mkdtemp(prefix="centurymetadata-"))
+    (basedir / "00-ff" / "00-ff").mkdir(parents=True, exist_ok=True)
+
+    httpd = CenturyHTTPServer(basedir, args.host, args.port)
+    port = httpd.server_address[1]
+    print(f"Serving centurymetadata API on http://{args.host}:{port}/api/v1/ (basedir={basedir})")
+    print("Ctrl-C to stop.")
+    try:
+        httpd.serve_forever()
+    except KeyboardInterrupt:
+        pass
+    finally:
+        if args.basedir is None:
+            shutil.rmtree(basedir, ignore_errors=True)
+
+
+if __name__ == "__main__":
+    main()
